Use view-based transition approach with shared elements
It kinda works but does not look similar to a transition between activities. One reason is that the fade transition keeps the _previous_ scene as an overlay which will be on top of the shared element. Second, the bounds transition is somehow different (doesn't look right) compared with the activity transition.